Create compile/mark/init for initializing orient for MarkDef
